requirements.txt
django-axes
celery
redis
django-celery-beat
django-celery-results